initiated controller to generate shopping list for users

+++ src/main/java/com/cookplanner/repositories/Recipes.java
@@ -9,5 +9,5 @@ import java.util.List;
 /**
  * @author Created by Yassine on 2/20/17.
  */
-public interface Recipes extends CrudRepository<Recipe, Integer> {
+public interface Recipes extends CrudRepository<Recipe, Long> {
 }
